restaurant-generator/requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai

python-dotenv
streamlit